refactor(shared): modularize gemini provider metadata
